+++ docs/changelog.rst
@@ -15,6 +15,7 @@ b. rm_anova, friedman and mixed_anova now support the presence of other repeated
 
 a. Added repeated measures correlation (Bakdash and Marusich 2017).
 b. Added robust skipped correlation (Rousselet and Pernet 2012).
+c. Pairwise_corr function now automatically delete non-numeric columns.
 
 **Dataset**